requirements.txt
pyttsx3
SpeechRecognition
transformers
langchain
PyAudio
soundfile
pydantic
pvporcupine
langchain_huggingface
python-dotenv
setuptools
torch
whisper
pathlib